refactor: payment 영속성 분리1

+++ refactoring_project/src/main/java/com/paymilli/paymilli/domain/payment/dto/response/PaymentResponse.java
@@ -12,7 +12,7 @@ public class PaymentResponse {
     private String cardName;
     private String cardImg;
     private int installment;
-    private int chargePrice;
+    private long chargePrice;
     private CardType cardType;
     private String approveNumber;
 }
